Change auto-install procedure to handle more edge cases
Some popular open source gems (e.g. octokit) cache their rack middleware
builder and reuse it across all connections. The first time this
middleware builder is used the handlers it contains will be frozen, and
can no longer be modified.

If the `AutoInstall` logic attempts to append the Honeycomb middleware
every time a connection is created then the app will error with an error
like:

```
Faraday::RackBuilder::StackLocked:
       can't modify middleware stack after making a request
```

One thing to note is that when Honeycomb's `AutoInstall` logic handles the
builder on the second connection the Honeycomb middleware is already
present in the builder, so there's no need to add it a second time.

The main change in this PR is that the `AutoInstall` logic will only
insert the middleware into the builder if it isn't already there.

Note that this will not prevent the `AutoInstall` logic erroring if a
locked middleware builder is passed to Faraday, and the Honeycomb
middleware is not present. This seems like more of an edge
case/something the developer might want to be aware of, so I'm not going
to handle it, but it should be quite easy to implement if needed.

The second major change is that the `AutoInstall` logic will always
attempt to add the middleware one level above the underlying HTTP
adapter. Some Faraday middleware (e.g. retry, follow redirects) will
cause Faraday to issue multiple HTTP requests in a single call.

In the previous code Honeycomb's middleware ran before all other
middleware, so it would miss these extra requests. By pushing the
middleware as close as possible to the adapter we should be able to
capture them.

Moving the middleware insertion logic out of the monkey patched method
allows us to collapse a lot of the if/else branches, which makes it a
lot easier to understand.

I've also tweaked the config parsing so that it handles the case where
the first argument to `Faraday.new` is a string URL.

Adding byebug/pry as a development dependency to make it easier to debug
the `AutoInstall` logic.

Fixes honeycombio/faraday-honeycomb#6