Do not raise error when propagating an incoming trace with context
The double splat operator only accepts hashes with symbol keys - if you
give it a hash with string keys it raises a runtime error. When the
active trace context was derived from an incoming HTTP header the keys
will contain strings, so ruby will raise a runtime error like:

```
wrong argument type String (expected Symbol)
(repl):4:in `<main>'
```

The `encode_trace_context` method only takes 3 arguments, so I'm not
sure why we need to try and turn the active trace context into keyword
arguments. The code now passes all context as a hash in the third
argument.

Unfortunately the codepath that adds the trace context header only runs
if the `honeycomb-beeline` gem is installed, so I had to add it as a
development dependency.

Fixes honeycombio/faraday-honeycomb#10